requirements.txt
fastapi 
pydantic
uvicorn 
python-jose[cryptography] 
passlib
python-multipart
numpy
matplotlib
starlette
typing-extensions
asgiref
clic
colorama
SQLAlchemy
psycopg2_binary
SQLAlchemy_utils
